chore: updated the profile setting page with new form
